refactor: split widgets into modules
