+++ crates/render/src/widget/image.rs
@@ -1,0 +1,120 @@
+use config::display::ImageProperty;
+use log::warn;
+
+use crate::{
+    drawer::Drawer,
+    image::Image,
+    types::{Offset, RectSize},
+};
+
+use super::{CompileState, Draw, WidgetConfiguration};
+
+#[derive(macros::GenericBuilder, Clone)]
+#[gbuilder(name(GBuilderWImage))]
+pub struct WImage {
+    #[gbuilder(hidden, default(Image::Unknown))]
+    content: Image,
+
+    #[gbuilder(hidden, default(0))]
+    width: usize,
+    #[gbuilder(hidden, default(0))]
+    height: usize,
+
+    #[gbuilder(default)]
+    property: ImageProperty,
+}
+
+impl WImage {
+    pub fn new() -> Self {
+        Self {
+            content: Image::Unknown,
+            width: 0,
+            height: 0,
+            property: Default::default(),
+        }
+    }
+
+    pub fn compile(
+        &mut self,
+        rect_size: RectSize,
+        WidgetConfiguration {
+            notification,
+            display_config,
+            override_properties,
+            ..
+        }: &WidgetConfiguration,
+    ) -> CompileState {
+        if *override_properties {
+            self.property = display_config.image.clone();
+        }
+
+        self.content = notification
+            .hints
+            .image_data
+            .as_ref()
+            .cloned()
+            .map(|image_data| Image::from_image_data(image_data, &self.property, &rect_size))
+            .or_else(|| {
+                notification
+                    .hints
+                    .image_path
+                    .as_deref()
+                    .or(Some(notification.app_icon.as_str()))
+                    .map(|svg_path| Image::from_svg(svg_path, &self.property, &rect_size))
+            })
+            .unwrap_or(Image::Unknown);
+
+        self.width = self
+            .content
+            .width()
+            .map(|width| width + self.property.margin.horizontal() as usize)
+            .unwrap_or(0);
+        self.height = self
+            .content
+            .height()
+            .map(|height| height + self.property.margin.vertical() as usize)
+            .unwrap_or(0);
+
+        if self.width > rect_size.width || self.height > rect_size.height {
+            warn!(
+                "The image doesn't fit to available space.\
+                \nThe image size: width={}, height={}.\
+                \nAvailable space: width={}, height={}.",
+                self.width, self.height, rect_size.width, rect_size.height
+            );
+            return CompileState::Failure;
+        }
+
+        if self.content.is_exists() {
+            CompileState::Success
+        } else {
+            CompileState::Failure
+        }
+    }
+
+    pub fn width(&self) -> usize {
+        self.width
+    }
+
+    pub fn height(&self) -> usize {
+        self.height
+    }
+}
+
+impl Default for WImage {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl Draw for WImage {
+    fn draw_with_offset(&self, offset: &Offset, drawer: &mut Drawer) {
+        if !self.content.is_exists() {
+            return;
+        }
+
+        // INFO: The ImageProperty initializes with Image so we can calmly unwrap
+        let offset = Offset::from(&self.property.margin) + offset.clone();
+        self.content.draw_with_offset(&offset, drawer);
+    }
+}
